requirements.txt
Flask
flask-cors
Werkzeug
pdfplumber
python-dotenv
google-generativeai
pymongo